+++ src/components/Input/Input.types.ts
@@ -1,10 +1,16 @@
 import { FormEvent } from 'react';
 
+export enum InputTypes {
+  Text = 'text',
+  Email = 'email',
+  Password = 'password',
+}
+
 export interface IInputProps {
   id: string;
   value: string;
   label: string;
-  type: 'text' | 'email' | 'password';
+  type: InputTypes;
   autoComplete?: 'on' | 'off';
   isValid?: boolean;
   isInvalid?: boolean;
