+++ src/pages/Profile/Profile.types.ts
@@ -1,10 +1,12 @@
+import { InputTypes } from '../../components/Input/Input.types';
+
 export interface IPropsProfile {
   className?: string;
 }
 
 export interface IProfileInput {
   id: string;
-  type: 'text' | 'email' | 'password';
+  type: InputTypes;
   label: string;
   controlName: ControlNames;
   className: string;
@@ -21,72 +23,3 @@ export enum ControlNames {
   NewPassword = 'new password',
   PasswordRepeat = 'password repeat',
 }
-
-export const PROFILE_INPUTS_DATA: IProfileInput[] = [
-  {
-    id: 'profile-email',
-    type: 'email',
-    label: 'Email',
-    controlName: ControlNames.Email,
-    className: 'email',
-  },
-  {
-    id: 'profile-login',
-    type: 'text',
-    label: 'Login',
-    controlName: ControlNames.Login,
-    className: 'login',
-  },
-  {
-    id: 'profile-first-name',
-    type: 'text',
-    label: 'First name',
-    controlName: ControlNames.FirstName,
-    className: 'first-name',
-  },
-  {
-    id: 'profile-last-name',
-    type: 'text',
-    label: 'Last name',
-    controlName: ControlNames.LastName,
-    className: 'last-name',
-  },
-  {
-    id: 'profile-display-name',
-    type: 'text',
-    label: 'Display name',
-    controlName: ControlNames.DisplayName,
-    className: 'display-name',
-  },
-  {
-    id: 'profile-phone',
-    type: 'text',
-    label: 'Phone',
-    controlName: ControlNames.Phone,
-    className: 'phone',
-  },
-];
-
-export const PROFILE_INPUTS_PASSWORD: IProfileInput[] = [
-  {
-    id: 'profile-new-pass',
-    type: 'password',
-    label: 'New password',
-    controlName: ControlNames.NewPassword,
-    className: 'new-pass',
-  },
-  {
-    id: 'profile-old-pass',
-    type: 'password',
-    label: 'Old password',
-    controlName: ControlNames.OldPassword,
-    className: 'old-pass',
-  },
-  {
-    id: 'profile-repeat-pass',
-    type: 'password',
-    label: 'Password (repeat)',
-    controlName: ControlNames.PasswordRepeat,
-    className: 'repeat-pass',
-  },
-];
