+++ src/services/hooks/useValidateControl/useValidateContol.types.ts
@@ -4,6 +4,9 @@ export interface IControlInfo {
   value: string;
   isValid: boolean;
   isInvalid: boolean;
+  /**
+   *. isDirty flag if the input has been changed
+   * */
   isDirty: boolean;
   errors: IValidationErrors;
   setValue: (value: string) => void;
