+++ src/services/index.ts
@@ -1,6 +1,6 @@
 export { http } from './axios';
 export { Validator } from './validatior';
 
-export { useValidateControl } from './hooks/useValidate';
+export { useValidateControl } from './hooks/useValidateControl';
 
 export * from './constants';
